Stop truncating MP3 bitrates

+++ tinytag/tinytag.py
@@ -601,7 +601,7 @@ class ID3(TinyTag):
             b = fh.peek(4)
             if len(b) < 4:
                 if frames:
-                    self.bitrate = int(bitrate_accu / frames)
+                    self.bitrate = bitrate_accu / frames
                 break  # EOF
             sync, conf, bitrate_freq, rest = struct.unpack('BBBB', b[0:4])
             br_id = (bitrate_freq >> 4) & 0x0F  # biterate id
@@ -633,7 +633,7 @@ class ID3(TinyTag):
                     xframes, byte_count, toc, vbr_scale = ID3._parse_xing_header(fh)
                     if xframes and xframes != 0 and byte_count:
                         self.duration = xframes * ID3.samples_per_frame / float(self.samplerate) / self.channels
-                        self.bitrate = int(byte_count * 8 / self.duration / 1000)
+                        self.bitrate = byte_count * 8 / self.duration / 1000
                         self.audio_offset = fh.tell()
                         return
                     continue
@@ -658,7 +658,7 @@ class ID3(TinyTag):
                 est_frame_count = audio_stream_size / (frame_size_accu / float(frames))
                 samples = est_frame_count * ID3.samples_per_frame
                 self.duration = samples / float(self.samplerate)
-                self.bitrate = int(bitrate_accu / frames)
+                self.bitrate = bitrate_accu / frames
                 return
 
             if frame_length > 1:  # jump over current frame body
